split event component
