+++ src/pages/index.js
@@ -8,19 +8,13 @@ import {
   darkAccent,
   styleLinks,
 } from '../components/styles';
-import ExternalLink from '../components/ExternalLink';
-
-const getDirectionsLink = (address) =>
-  `https://www.google.com/maps/dir/?api=1&destination=${encodeURIComponent(
-    address
-  )}`;
+import EventPreview from '../components/EventPreview';
 
 function HomePage({
   data: {
     allEventbriteEvent: { edges: events },
   },
 }) {
-  console.log(events);
   return (
     <Layout>
       <div className="page-container">
@@ -49,32 +43,13 @@ function HomePage({
         </div>
         <div css={[withGutters, styleLinks({ color: darkAccent })]}>
           <h1 css={{ margin: '0.75em 0' }}>Upcoming Events</h1>
-          {events.map(({ node: event }) => {
-            const displayAddress =
-              event.venue.address.localized_address_display;
-            return (
-              <article css={{ marginBottom: '1rem' }} key={event.id}>
-                <h3>
-                  <ExternalLink href={event.url}>
-                    {event.name.text}
-                  </ExternalLink>
-                </h3>
-                <p>
-                  <span>{event.venue.name}</span>
-                  <br />
-                  {displayAddress}
-                  <br />
-                  <ExternalLink href={event.url}>
-                    RSVP
-                  </ExternalLink>{' '}
-                  |{' '}
-                  <ExternalLink href={getDirectionsLink(displayAddress)}>
-                    Get Directions
-                  </ExternalLink>
-                </p>
-              </article>
-            );
-          })}
+          {events.map(({ node: event }) => (
+            <EventPreview
+              key={event.id}
+              event={event}
+              css={{ marginBottom: '1rem' }}
+            />
+          ))}
         </div>
         {/*
         <div css={[withGutters]}>
